# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
  account.accountResponse:
    properties:
      balance:
        type: integer
      cpf:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  account.balanceResponse:
    properties:
      balance:
        type: integer
    type: object
  account.createdAccountResponse:
    properties:
      balance:
        type: integer
      cpf:
        type: string
      name:
        type: string
    type: object
  account.requestBody:
    properties:
      cpf:
        type: string
      name:
        type: string
      secret:
        type: string
    type: object
  auth.requestBody:
    properties:
      cpf:
        type: string
      secret:
        type: string
    type: object
  auth.responseBody:
    properties:
      token:
        type: string
    type: object
  transfer.transferCreatedResponse:
    properties:
      account_destination_id:
        type: string
      account_origin_id:
        type: string
      amount:
        type: integer
    type: object
  transfer.transferRequest:
    properties:
      account_destination_id:
        type: string
      amount:
        type: integer
    type: object
  transfer.transferResponse:
    properties:
      account_destination_id:
        type: string
      amount:
        type: integer
      created_at:
        type: string
    type: object
host: localhost:5000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a simple banking api.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Banking API
  version: "1.0"
paths:
  /accounts:
    get:
      consumes:
      - application/json
      description: List all accounts.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/account.accountResponse'
            type: array
        "500":
          description: Internal server error
          schema:
            type: string
      summary: List accounts
      tags:
      - accounts
    post:
      consumes:
      - application/json
      description: |-
        Creates a new account given a name, cpf and secret.
        Secret must be a minimum of 8, a maximum of 20, at least one lowercase character,
        one uppercase character and one number.
        CPF must have the format XXX.XXX.XXX-XX or XXXXXXXXXXX.
      parameters:
      - description: Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/account.requestBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/account.createdAccountResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "409":
          description: Account already exists
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Creates a new account
      tags:
      - accounts
  /accounts/{accountID}/balance:
    get:
      consumes:
      - application/json
      description: Get account balance by accountID, if exists.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/account.balanceResponse'
        "400":
          description: Account does not exist
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get account balance
      tags:
      - accounts
  /login:
    post:
      consumes:
      - application/json
      description: Returns a JWT to be used on /transfers endpoints.
      parameters:
      - description: Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/auth.requestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.responseBody'
        "400":
          description: Bad request
          schema:
            type: string
        "404":
          description: Account does not exist
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Log in
      tags:
      - login
  /transfers:
    get:
      consumes:
      - application/json
      description: List all transfers.
      parameters:
      - description: Bearer Authorization Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/transfer.transferResponse'
            type: array
        "401":
          description: No token provided
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: List transfers
      tags:
      - transfers
    post:
      consumes:
      - application/json
      description: Creates a new transfer between two accounts.
      parameters:
      - description: Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/transfer.transferRequest'
      - description: Bearer Authorization Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/transfer.transferCreatedResponse'
            type: array
        "401":
          description: No token provided
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Create a new transfer
      tags:
      - transfers
swagger: "2.0"